+++ samples/Directory.Build.targets
@@ -1,0 +1,30 @@
+﻿<Project ToolsVersion="15.0">
+	<!--
+	This is a temporary workaround to avoid error "NETSDK1152: Found multiple publish output files with the same relative path:"
+	for Microsoft.Web.WebView2.Core.dll, with one coming from MsixContent and the other from the Microsoft.Web.Webview2 Nuget package.
+	If both are present, we only keep the one from the NuGet package. See https://github.com/unoplatform/uno/issues/14555.
+	-->
+	<Target Name="ResolveWebView2CoreDuplicates1" BeforeTargets="_ComputeResolvedFilesToPublishTypes" AfterTargets="ComputeFilesToPublish">
+		<Message Importance="high" Text ="Applying workaround to resolve Microsoft.Web.WebView2.Core.dll duplication in package (1)" />
+		<ItemGroup>
+			<_WebView2CoreFilesToExclude Include="@(ResolvedFileToPublish)" Condition="'%(Filename)' == 'Microsoft.Web.WebView2.Core'"/>
+		</ItemGroup>
+		<ItemGroup Condition="'@(_WebView2CoreFilesToExclude->Count())' &gt; 1">
+			<_WebView2CoreFilesToExclude Remove="@(_WebView2CoreFilesToExclude)" Condition="$([System.String]::Copy(%(FullPath)).Contains('.nuget'))"/>
+			<ResolvedFileToPublish Remove="@(_WebView2CoreFilesToExclude)" />
+		</ItemGroup>
+		<Message Importance="high" Text ="Removed: @(_WebView2CoreFilesToExclude)" />
+	</Target>
+
+	<Target Name="ResolveWebView2CoreDuplicates2" BeforeTargets="_ComputeAppxPackagePayload" AfterTargets="GetPackagingOutputs">
+		<Message Importance="high" Text ="Applying workaround to resolve Microsoft.Web.WebView2.Core.dll duplication in package (2)" />
+		<ItemGroup >
+			<_WebView2CoreOutputsToExclude Include="@(PackagingOutputs)" Condition="'%(Filename)' == 'Microsoft.Web.WebView2.Core'"/>
+		</ItemGroup>
+		<ItemGroup Condition="'@(_WebView2CoreOutputsToExclude->Count())' &gt; 1">
+			<_WebView2CoreOutputsToExclude Remove="@(_WebView2CoreOutputsToExclude)" Condition="$([System.String]::Copy(%(FullPath)).Contains('.nuget'))"/>
+			<PackagingOutputs Remove="@(_WebView2CoreOutputsToExclude)" />
+		</ItemGroup>
+		<Message Importance="high" Text ="Removed: @(_WebView2CoreOutputsToExclude)" />
+	</Target>
+</Project>
