+++ samples/MauiEmbedding/Directory.Build.targets
@@ -1,2 +1,3 @@
-﻿<Project>
+﻿<Project ToolsVersion="15.0">
+	<Import Project="$(MSBuildThisFileDirectory)..\Directory.Build.targets" />
 </Project>
