+++ samples/Playground/Playground/Platforms/Desktop/Program.cs
@@ -1,4 +1,5 @@
 using Uno.UI.Runtime.Skia;
+using Uno.UI.Hosting;
 
 namespace Playground;
 public class Program
@@ -6,12 +7,12 @@ public class Program
     [STAThread]
     public static void Main(string[] args)
     {
-        var host = SkiaHostBuilder.Create()
+        var host = UnoPlatformHostBuilder.Create()
             .App(() => new App())
             .UseX11()
             .UseLinuxFrameBuffer()
             .UseMacOS()
-            .UseWindows()
+            .UseWin32()
             .Build();
 
         host.Run();
