+++ src/Directory.Build.props
@@ -15,7 +15,6 @@
 		<!-- https://github.com/unoplatform/private/issues/332 -->
 		<!-- <Nullable>annotations</Nullable> -->
 		<Nullable>enable</Nullable>
-		
 		<TreatWarningsAsErrors>true</TreatWarningsAsErrors>
 		<NoWarn>$(NoWarn);NU5104;NU5100;NU5118;NU5123;NU1603;CS1701;CS1702;XA0101;MSB3277;CS8785;CS8669;CS1998;NU1507</NoWarn>
 
@@ -28,10 +27,9 @@
 		<GenerateLibraryLayout>true</GenerateLibraryLayout>
 		<ImplicitUsings>true</ImplicitUsings>
 		<OutputType>Library</OutputType>
-
 		<UnoExtensionsGeneration_DisableCreateNewOnMetadataUpdateAttribute>True</UnoExtensionsGeneration_DisableCreateNewOnMetadataUpdateAttribute>
 
-
+		<WindowsSdkPackageVersion>10.0.19041.57</WindowsSdkPackageVersion>
 		<!-- Enables hard links for obj -> bin files (symlinks generally don't work there -->
 		<CreateHardLinksForCopyFilesToOutputDirectoryIfPossible>true</CreateHardLinksForCopyFilesToOutputDirectoryIfPossible>
