+++ src/Uno.Extensions.Maui.UI/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using global::System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Uno.Extensions.Maui.WinUI.Runtime.Skia")]
