+++ src/Uno.Extensions.Maui.UI/Internals/IPlatformViewProvider.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Uno.Extensions.Maui.Internals
+{
+	internal interface IPlatformViewProvider
+	{
+		object? GetPlatformView(IMauiContext mauiContext);
+	}
+}
