+++ src/Uno.Extensions.Maui.UI/MauiEmbedding.cs
@@ -2,6 +2,8 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Maui;
 using Microsoft.Maui.ApplicationModel;
 using Uno.Extensions.Hosting;
+using Uno.Extensions.Maui.Extensibility;
+using Uno.Extensions.Maui.Platform;
 
 namespace Uno.Extensions.Maui;
 
@@ -82,76 +84,8 @@ public static partial class MauiEmbedding
 
 	private static MauiApp BuildMauiApp(MauiAppBuilder builder, Application app, Microsoft.UI.Xaml.Window window)
 	{
-		var mauiApp = builder.Build();
-		mauiApp.InitializeMauiEmbeddingApp(app);
-
-#if WINDOWS
-		window.Activated += (s, args) =>
-		{
-			WindowStateManager.Default.OnActivated(window, args);
-		};
-#endif
-		return mauiApp;
+		return MauiEmbeddingExtension.Default.BuildMauiApp(builder, app, window);
 	}
-
-	private static void InitializeScopedServices(this IMauiContext scopedContext)
-	{
-		var scopedServices = scopedContext.Services.GetServices<IMauiInitializeScopedService>();
-
-		foreach (var service in scopedServices)
-		{
-			service.Initialize(scopedContext.Services);
-		}
-	}
-
-	private static void InitializeApplicationMainPage(IApplication iApp)
-	{
-		if (iApp is not MauiApplication app || app.Handler?.MauiContext is null)
-		{
-			// NOTE: This method is supposed to be called immediately after we initialize the Application Handler
-			// This should never actually happen but is required due to nullability
-			return;
-		}
-
-#if ANDROID
-		var services = app.Handler.MauiContext.Services;
-		var context = new MauiContext(services, services.GetRequiredService<Android.App.Activity>());
-#else
-		var context = app.Handler.MauiContext;
-#endif
-
-		// Create an Application Main Page and initialize a Handler with the Maui Context
-		var page = new ContentPage();
-		app.MainPage = page;
-		_ = page.ToPlatform(context);
-
-		// Create a Maui Window and initialize a Handler shim. This will expose the actual Application Window
-		var virtualWindow = new Microsoft.Maui.Controls.Window();
-		virtualWindow.Handler = new EmbeddedWindowHandler
-		{
-#if IOS || MACCATALYST
-			PlatformView = context.Services.GetRequiredService<UIKit.UIWindow>(),
-#elif ANDROID
-			PlatformView = context.Services.GetRequiredService<Android.App.Activity>(),
-#elif WINDOWS
-			PlatformView = context.Services.GetRequiredService<Microsoft.UI.Xaml.Window>(),
-#endif
-			VirtualView = virtualWindow,
-			MauiContext = context
-		};
-		virtualWindow.Page = page;
-
-		app.SetCoreWindow(virtualWindow);
-	}
-
-	private static void SetCoreWindow(this IApplication app, Microsoft.Maui.Controls.Window window)
-	{
-		if (app.Windows is List<Microsoft.Maui.Controls.Window> windows)
-		{
-			windows.Add(window);
-		}
-	}
-
 #endif
 
 	internal record EmbeddedApplication : IPlatformApplication
@@ -165,6 +99,12 @@ public static partial class MauiEmbedding
 
 		public IServiceProvider Services { get; }
 		public IApplication Application { get; }
+	}
+
+	private static MauiAppBuilder RegisterPlatformServices(this MauiAppBuilder builder, Application app)
+	{
+		MauiEmbeddingExtension.Default.RegisterPlatformServices(builder, app);
+		return builder;
 	}
 
 	// NOTE: This was part of the POC and is out of scope for the MVP. Keeping it in case we want to add it back later.
