+++ src/Uno.Extensions.Maui.UI/PlatformViewProvider.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Uno.Extensions.Maui
+{
+	internal class PlatformViewProvider
+	{
+	}
+}
