+++ src/Uno.Extensions.RuntimeTests/Uno.Extensions.RuntimeTests/Platforms/Desktop/Program.cs
@@ -1,4 +1,5 @@
 using Uno.UI.Runtime.Skia;
+using Uno.UI.Hosting;
 
 namespace Uno.Extensions.RuntimeTests;
 public class Program
@@ -8,12 +9,12 @@ public class Program
     {
         App.InitializeLogging();
 
-        var host = SkiaHostBuilder.Create()
+        var host = UnoPlatformHostBuilder.Create()
             .App(() => new App())
             .UseX11()
             .UseLinuxFrameBuffer()
             .UseMacOS()
-            .UseWindows()
+            .UseWin32()
             .Build();
 
         host.Run();
