+++ src/tfms-ui-maui.props
@@ -11,6 +11,6 @@
 		<TargetFrameworks>$(TargetFrameworks);net8.0</TargetFrameworks>
 	</PropertyGroup>
 	<PropertyGroup>
-		<WindowsSdkPackageVersion>10.0.19041.38</WindowsSdkPackageVersion>
+		<WindowsSdkPackageVersion>10.0.19041.57</WindowsSdkPackageVersion>
 	</PropertyGroup>
 </Project>
