+++ testing/TestHarness/TestHarness/Platforms/Desktop/Program.cs
@@ -1,3 +1,4 @@
+using Uno.UI.Hosting;
 using Uno.UI.Runtime.Skia;
 
 namespace TestHarness;
@@ -7,12 +8,12 @@ public class Program
 	public static void Main(string[] args)
 	{
 
-		var host = SkiaHostBuilder.Create()
+		var host = UnoPlatformHostBuilder.Create()
 			.App(() => new App())
 			.UseX11()
 			.UseLinuxFrameBuffer()
 			.UseMacOS()
-			.UseWindows()
+			.UseWin32()
 			.Build();
 
 		host.Run();
